Port menu_modificar unix

+++ src/menu.h
@@ -19,7 +19,7 @@ void menu_item(char *string, char *color1, char *color2);
 void menu_centered_item(char *string, char *color1, char *color2,
                         int32_t row_offset);
 void clear_menu(void);
-int32_t arrow_menu(char *strings[], int32_t size);
+int arrow_menu(char *strings[], size_t size);
 void cursor_upLeft(void);
 void hide_cursor(void);
 void show_cursor(void);
@@ -29,7 +29,7 @@ void disableRawMode(void);
 #endif
 
 void menu_principal(void);
-int32_t input_menu(Input inputItems[], int32_t inputItemsSize, int32_t isVenda);
+int input_menu(Input inputItems[], int inputItemsSize, int isVenda);
 void printMenuItem(Input item, int32_t isSelected, int32_t offset);
 void printMenuCheckbox(char *string, int32_t isSelected, int32_t isCheckboxSelected, int32_t offset);
 int arrow_menu_filter(Livro *livros, int size);
